+++ components/led/led_indicator/src/led_rgb.c
@@ -21,9 +21,10 @@
     }
 
 typedef struct {
+    bool is_active_level_high;     /*!< Set true if GPIO level is high when light is ON, otherwise false. */
     ledc_channel_t rgb_channel[3]; /*!< RGB Channels */
     uint32_t max_duty;             /*!< Max duty cycle from duty_resolution : 2^duty_resolution -1 */
-    uint32_t hsv;                  /*!< HSV: H [0-360] - 9 bits, S [0-255] - 8 bits, V [0-255] - 8 bits*/
+    led_indicator_ihsv_t hsv;      /*!< HSV: H [0-360] - 9 bits, S [0-255] - 8 bits, V [0-255] - 8 bits*/
 } led_rgb_t;
 
 esp_err_t led_indicator_rgb_init(void *param, void **ret_rgb)
@@ -51,6 +52,7 @@ esp_err_t led_indicator_rgb_init(void *param, void **ret_rgb)
     rgb->rgb_channel[0] = cfg->red_channel;
     rgb->rgb_channel[1] = cfg->green_channel;
     rgb->rgb_channel[2] = cfg->blue_channel;
+    rgb->is_active_level_high = cfg->is_active_level_high;
     *ret_rgb = (void *)rgb;
     return ESP_OK;
 EXIT:
@@ -70,6 +72,11 @@ esp_err_t led_indicator_rgb_deinit(void *rgb_handle)
 static esp_err_t led_indicator_rgb_set_duty(led_rgb_t *p_rgb ,uint32_t rgb[])
 {
     esp_err_t ret;
+    if (!p_rgb->is_active_level_high) {
+        rgb[0] = p_rgb->max_duty - rgb[0];
+        rgb[1] = p_rgb->max_duty - rgb[1];
+        rgb[2] = p_rgb->max_duty - rgb[2];
+    }
     for (int i = 0; i < 3; i++) {
         ret = ledc_set_duty(LEDC_MODE, p_rgb->rgb_channel[i], rgb[i]);
         LED_RGB_CHECK(ESP_OK == ret, "LEDC set duty error", return ret);
@@ -86,7 +93,8 @@ esp_err_t led_indicator_rgb_set_on_off(void *rgb_handle, bool on_off)
     led_rgb_t *p_rgb = (led_rgb_t *)rgb_handle;
 
     if (on_off) {
-        uint32_t rgb[3] = {p_rgb->max_duty, p_rgb->max_duty, p_rgb->max_duty};
+        uint32_t rgb[3] = {0};
+        led_indicator_hsv2rgb(p_rgb->hsv.value, &rgb[0], &rgb[1], &rgb[2]);
         ret = led_indicator_rgb_set_duty(p_rgb, rgb);
         LED_RGB_CHECK(ESP_OK == ret, "LEDC set duty error", return ret);
     } else {
@@ -109,7 +117,7 @@ esp_err_t led_indicator_rgb_set_rgb(void *rgb_handle, uint32_t rgb_value)
 
     ret = led_indicator_rgb_set_duty(p_rgb, rgb);
     LED_RGB_CHECK(ESP_OK == ret, "LEDC set duty error", return ret);
-    p_rgb->hsv = led_indicator_rgb2hsv(rgb_value);
+    p_rgb->hsv.value = led_indicator_rgb2hsv(rgb_value);
     return ESP_OK;
 }
 
@@ -125,7 +133,7 @@ esp_err_t led_indicator_rgb_set_hsv(void *rgb_handle, uint32_t hsv_value)
 
     ret = led_indicator_rgb_set_duty(p_rgb, rgb);
     LED_RGB_CHECK(ESP_OK == ret, "LEDC set duty error", return ret);
-    p_rgb->hsv = hsv_value;
+    p_rgb->hsv.value = hsv_value;
     return ESP_OK;
 }
 
@@ -134,8 +142,8 @@ esp_err_t led_indicator_rgb_set_brightness(void *rgb_handle, uint32_t brightness
     esp_err_t ret;
     led_rgb_t *p_rgb = (led_rgb_t *)rgb_handle;
     uint32_t rgb[3];
-    SET_BRIGHTNESS(p_rgb->hsv, brightness);
-    led_indicator_hsv2rgb(p_rgb->hsv, &rgb[0], &rgb[1], &rgb[2]);
+    p_rgb->hsv.v = brightness;
+    led_indicator_hsv2rgb(p_rgb->hsv.value, &rgb[0], &rgb[1], &rgb[2]);
     for (int i = 0; i < 3; i++) {
        rgb[i] = rgb[i] * p_rgb->max_duty / UINT8_MAX;
     }